Rewrote README to be clearer
Dropped all reference to how this is a game because it consistently confused students. Instead it's been turned into a tracking system that just keeps track of how many times a specific buzzword was used in a conversation. The "system" can be loaded with up to 5 buzzwords. The "PUT" method is used as a way to mark that a given buzzword was heard and the points should be updated accordingly. The "PUT" response should be the buzzword itself along with the new total score. All other aspects of the exercise was not modified.